Fixed Speed Perfomance of BTAllocator & Fix Free Space Calculation & Renew The Benchmarks in README

+++ inc/border_tags_allocator.h
@@ -20,8 +20,8 @@ typedef struct SHeap
     void*        m_buffer;
     BlockHeader* m_firstBlock;
     BlockFooter* m_lastFooter;
-    int          m_bufferSize;
-    int          m_freeSpace;
+    size_t       m_bufferSize;
+    size_t       m_freeSpace;
 } BTagsHeap;
 
 void  setupBTagsAllocator(void* buf, size_t size, BTagsHeap* heap);
